fixed concave mesh generation and fixed shader includes getting ignored when indented

+++ GameEngine/include/GameEngine/Rendering/VertexBuffer.h
@@ -11,7 +11,7 @@ namespace GameEngine
         class VertexBuffer : public Buffer<GL_ARRAY_BUFFER>
         {
             public:
-                VertexBuffer(const unsigned char* buffer, const unsigned elementSize, const size_t numElements, const GLenum drawType = GL_STATIC_DRAW):
+                VertexBuffer(const unsigned char* buffer, const size_t elementSize, const size_t numElements, const GLenum drawType = GL_STATIC_DRAW):
                     Buffer(buffer, elementSize, numElements, drawType) {}
         };
     }
